src/app2-modular/requirements.txt
# Core requirements for app2-modular
python-dotenv
azure-identity
semantic-kernel
